Add new game ui
This part is pretty cool because nothing is hard-coded.  Everything is
generated from the hyper-schemas that are attached to the API.  If I
change a schema, the UI updates automatically.

I used alpaca to generate the form from the create.json schema.

I used Jsonary to read and process hyper-schemas.